Update crash trial

+++ crash_trial.sh
@@ -1,16 +1,29 @@
-BEST_STEPS=9999
 TEMP_DIR=${1}_TMP
-for (( SPEED=30; SPEED<=100; SPEED++ ))
+START=${2}
+COUNT=${3}
+MAX_STEPS=${4}
+BEST_STEPS=$MAX_STEPS
+
+
+for (( SPEED=$START; SPEED<$START + $COUNT; SPEED++ ))
 do
-    rm -r $TEMP_DIR
-    python3.7 examples/bugtest.py scenarios/loop --speed ${SPEED} --save-dir $TEMP_DIR --write
+    if [[ -d $TEMP_DIR ]]; then
+        rm -r $TEMP_DIR
+    fi
+    PYTHONHASHSEED=42 python3.7 examples/bugtest.py scenarios/loop --speed ${SPEED} --save-dir $TEMP_DIR --max-steps $MAX_STEPS --write --headless
     STEPS=$?
     if (( $STEPS < $BEST_STEPS )); then
         echo $STEPS > ${TEMP_DIR}/steps.txt
         echo $SPEED > ${TEMP_DIR}/speed.txt
         echo "Steps taken ${STEPS}0, Speed: ${SPEED}"
-        rm -r $1
+        if [[ -d $1 ]]; then
+            rm -r $1
+        fi
         mv ${TEMP_DIR} $1
         BEST_STEPS=$STEPS
+        break
     fi
 done
+if [[ -d $TEMP_DIR ]]; then
+    rm -r $TEMP_DIR
+fi
